fix(fas2): åtgärda Fas 2-kodgranskningens verifierade fynd
HIGH:
- ::int-cast på bigint-ören-kolumner kapade alla belopp vid 2 147 483 647 ören
  (~21,47 MSEK) → 500. Ersatt med en pg-typparser (int8 → JS-tal, säkert eftersom
  belopp asserteras ≤ 2^53 vid skrivning). Gällde även Fas 1:s verifikatläsning.

MEDIUM:
- fillagring skriver nu till disk UTANFÖR transaktionen med städning vid rollback
  (faktura-PDF + kvittobilaga) — höll förut en poolklient under disk-I/O och
  lämnade föräldralösa filer; matchar files.ts-mönstret
- kvantitet avrundas till 3 decimaler (numeric(14,3)) så 0.0004 avvisas rent
  (400) i stället för CHECK-brott → 500; lagrat antal = beräknat antal
- nollbelopp: kvitto kräver positivt netto; faktura filtrerar nollrader vid
  bokföring (en gratisrad blockerar inte längre hela bokföringen)

LOW/härdning:
- composite-FK (migration 0011) binder faktura→kund och kvitto→leverantör till
  bolaget i DB (matchar Fas 1-mönstret)
- annullerade dokument kan inte bokföras; förfallodatum ≥ fakturadatum
- tidig kontovalidering (fail-fast) i faktura/kvitto; assertAccountsExist
  centraliserad i accounts.ts
- delad IsoDateSchema (dedup av 3 kopior); bolagsmejl via EmailSchema

8 nya regressionstester. Alla 103 tester passerar, typecheck rent.

+++ server/src/services/parties.ts
@@ -121,7 +121,7 @@ export async function updateSupplier(
 }
 
 // ---- Artiklar ----
-const ARTICLE_COLUMNS = `id, article_number, name, unit, unit_price_ore::int, vat_rate, revenue_account, is_active, created_at`;
+const ARTICLE_COLUMNS = `id, article_number, name, unit, unit_price_ore, vat_rate, revenue_account, is_active, created_at`;
 const ARTICLE_UPDATE = {
   name: 'name', unit: 'unit', unit_price_ore: 'unit_price_ore', vat_rate: 'vat_rate',
   revenue_account: 'revenue_account', is_active: 'is_active',
